lp-gateway: Move message processing logic to a new file
